requirements.txt
matplotlib==3.9.4
numpy==1.26.4
openai==1.88.0
pandas==2.2.2
plotly==6.1.2
python-dotenv==1.1.0
scikit-learn==1.5.0
streamlit==1.45.1
